Update Attribute.js
There was no case in the code for when useDocumentTypes === true, and from what I can tell from https://dynamoosejs.com/api#attribute-types this should be the proper code. I'm not sure exactly what debug() text should be included, so I went without. Unfortunately this crashes for me, so there must be more to it.

Intended to fix #230, which I am currently experiencing as well.